+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # @oslojs/jwt
+
+## 0.2.0
+
+- [Breaking] Rename `JWSRegisteredHeaderParameters` to `JWSRegisteredHeaders`
+- [Breaking] Update `createJWT()`, `parseJWT()` and `createJWTSignatureMessage()`
+- Add `decodeJWT()`